+++ tests/test_hive_client.py
@@ -1,34 +1,28 @@
-from api_client.python.client import HiveClient
+from python.client import HiveClient
 import random
 import pandas as pd
 
 ### Local Testing
-# Run app.py locally
-# comment these following lines
-# response = api_client.verify(auth_token)
-# user_id = response["sub"]
+# Run app.py locally and use the token from the console to test the client
 
 if __name__ == "__main__":
-    client = HiveClient(token="eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiJuYlowNUVqS2ZqWGpXdDBTMDdvOSIsImF1ZCI6ImFwcC5jcnlwdGljb3JuLmNvbSIsImlzcyI6ImFjY291bnRzLmNyeXB0aWNvcm4uY29tIiwianRpIjoicTdQMVh3WG9MdU5wbEJJNlFMVWMiLCJpYXQiOjE3Mjg4MTk1MDAsImV4cCI6MTcyODgyMzEwMCwic2NvcGVzIjpbInJlYWQ6cHJlZGljdGlvbnMiXX0.zbV9fQUmIQ1yYwaaAHazc1eYm6kK9-YMorOcWyWiuYg")
+    client = HiveClient(token="eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiJuYlowNUVqS2ZqWGpXdDBTMDdvOSIsImF1ZCI6ImFwcC5jcnlwdGljb3JuLmNvbSIsImlzcyI6ImFjY291bnRzLmNyeXB0aWNvcm4uY29tIiwianRpIjoiamhXVXVDS0pNVjQ0ZmNiWEx4VmMiLCJpYXQiOjE3Mjg5OTc4ODksImV4cCI6MTcyOTAwMTQ4OSwic2NvcGVzIjpbInJlYWQ6cHJlZGljdGlvbnMiXX0.oeJqCJ63kikLEgv83wrnicnBapz3dFujAmI5D3oGeTQ")
     
     print("Testing create_account:")
     #print(client.create_account())
-
-    #
-    # for i in range(10):
-    # #
+    
     print("\nTesting create_model:")
-    #print(client.create_model(1, "Tatooine"))
+    #print(client.create_model(1, "Alderaan"))
     #     #
-    #     print("\nTesting evaluate_model:")
-    #     data = pd.DataFrame(columns=["feature1"], data=[random.gauss(1,0.02) for _ in range(4000)])
-    #     print(client.evaluate_model(i, data))
     print("\nTesting evaluate_model:")
     #data = pd.DataFrame(columns=["feature1"], data=[random.gauss(1,0.02) for _ in range(4986)])
     #print(client.evaluate_model(1, data))
 
     print("\nTesting get_specific_model:")
-    #print(client.get_model(2))
+    #print(client.get_model(id=1))
+    
+    print("\nTesting get_all_models:")
+    #print(client.get_model())
 
     print("\nTesting get_leaderboard:")
     #print(client.get_leaderboard())
@@ -48,5 +42,11 @@ if __name__ == "__main__":
     print("\nTesting download_data:")
     #print(client.download_data(model_id=1, feature_size="small", version=1))
 
-    print("\nTesting get_account_info:")
-    #print(client.get_account_info(username="Mister Y"))
+    print("\nTesting get_account_info by id:")
+    #print(client.get_account_info(id="nbZ05EjKfjXjWt0S07o9"))
+    
+    print("\nTesting get_account_info by username:")
+    #print(client.get_account_info(id="mateh"))
+    
+    print("\nTesting get_account_info (current user):")
+    #print(client.get_account_info())