+++ public/CHANGELOG.md
@@ -8,7 +8,6 @@
 ###### 13.10.2024
 - Bug fixes
 - Renaming data info endpoint
-- Add benchmarks to model evaluation
 - Remove help endpoint
 
 ### v0.2.1
@@ -16,3 +15,9 @@
 - Account creation and account retrieval defaults to logged in user
 - Return model id in account response
 - Remove api generation message
+
+### v0.3.0
+###### 15.10.2024
+- Account retrieval with username
+- Retrieve all models
+- Add benchmarks to model evaluation